# unity_sps_ogc_processes_api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: info@ogc.org
    name: Open Geospatial Consortium
  description: Example API Definition for OGC API - Processes
  license:
    name: OGC License
    url: http://www.opengeospatial.org/legal/
  title: OGC API - Processes
  version: "0.1"
servers:
- description: "Example Server 1 (GNOSIS, supporting Part 1: Core /Sync and Part 3:\
    \ Workflows)"
  url: https://maps.gnosis.earth/ogcapi
paths:
...
  /processes/{processId}:
    get:
      description: |
        The process description contains information about inputs and outputs and a link to the execution-endpoint for the process. The Core does not mandate the use of a specific process description to specify the interface of a process. That said, the Core requirements class makes the following recommendation:

        Implementations SHOULD consider supporting the OGC process description.

        For more information, see [Section 7.8](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_description).
      operationId: getProcessDescription
      parameters:
      - in: path
        name: processId
        required: true
        schema:
          $ref: '#/components/schemas/processes-list'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/process'
          description: A process description.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
      summary: retrieve a process description
      tags:
      - Processes
  /processes/{processId}/execution:
    post:
      callbacks:
        jobCompleted:
          '{$request.body#/subscriber/successUri}':
            post:
              operationId: jobCompleted_Post
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/results_1'
              responses:
                "200":
                  description: Results received successfully
              x-callback-request: true
      description: |
        Executes a process (this may result in the creation of a job resource e.g., for _asynchronous execution_).

        For more information, see [Section 7.9](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_create_job).
      operationId: execute
      parameters:
      - in: path
        name: processId
        required: true
        schema:
          $ref: '#/components/schemas/processes-list'
      - description: |-
          Indicates client preferences, including whether the client is capable of asynchronous processing.
          A `respond-async` preference indicates a preference for asynchronous processing.
          A `wait: <x>s` preference indicates that the client prefers to wait up to x seconds to receive a reponse synchronously before the server falls back to asynchronous processing.
        in: header
        name: Prefer
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/execute'
        description: |-
          An execution request specifying any inputs for the process to execute, and optionally to select specific outputs.
          With support for _Part 3: Workflows and chaining_, this execution request may specify a complex processing workflow
          e.g., including nested processes and OGC API collections as inputs.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/execute_200_response'
            image/png:
              schema:
                format: binary
                type: string
            image/jpeg:
              schema:
                format: binary
                type: string
            image/tiff; application=geotiff:
              schema:
                format: binary
                type: string
            application/geo+json:
              schema:
                $ref: '#/components/schemas/execute_200_response_1'
          description: Result of synchronous execution
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/statusInfo'
          description: Started asynchronous execution. Created job.
          headers:
            Location:
              description: URL to check the status of the execution/job.
              schema:
                type: string
            Preference-Applied:
              description: The preference applied to execute the process asynchronously
                (see. RFC 2740).
              schema:
                type: string
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: execute a process.
      tags:
      - Processes
  /jobs:
    get:
      description: |
        Lists available jobs.

        For more information, see [Section 12](https://docs.ogc.org/is/18-062r2/18-062r2.html#Job_list).
      operationId: getJobs
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobList'
          description: A list of jobs for this process.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
      summary: retrieve the list of jobs.
      tags:
      - Jobs
  /jobs/{jobId}:
    delete:
      description: |
        Cancel a job execution and remove it from the jobs list.

        For more information, see [Section 14]https://docs.ogc.org/is/18-062r2/18-062r2.html#Dismiss).
      operationId: dismiss
      parameters:
      - description: local identifier of a job
        in: path
        name: jobId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/statusInfo'
          description: The status of a job.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: "cancel a job execution, remove a finished job"
      tags:
      - Jobs
    get:
      description: |
        Shows the status of a job.

         For more information, see [Section 7.10](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_retrieve_status_info).
      operationId: getStatus
      parameters:
      - description: local identifier of a job
        in: path
        name: jobId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/statusInfo'
          description: The status of a job.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: retrieve the status of a job
      tags:
      - Jobs
  /jobs/{jobId}/results:
    get:
      description: |
        Lists available results of a job. In case of a failure, lists exceptions instead.

        For more information, see [Section 7.11](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_retrieve_job_results).
      operationId: getResult
      parameters:
      - description: local identifier of a job
        in: path
        name: jobId
        required: true
        schema:
          type: string
      - description: |-
          Indicates client preferences, such as whether the client wishes a self-contained or minimal response.
          A `return=minimal` preference indicates that the client would prefer that links be returned to larger object to minimize the response payload.
          A `return=representation` indicates that the client would prefer if the server can return a self-contained response.
        in: header
        name: Prefer
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/results'
          description: The processing results of a job.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: retrieve the result(s) of a job
      tags:
      - Jobs
components:
...
  schemas:
    processSummary:
      allOf:
      - $ref: '#/components/schemas/descriptionType'
      - properties:
          id:
            type: string
          version:
            type: string
          jobControlOptions:
            items:
              $ref: '#/components/schemas/jobControlOptions'
            type: array
          links:
            items:
              $ref: '#/components/schemas/link'
            type: array
        required:
        - id
        - version
        type: object
      example:
        metadata:
        - hreflang: en
          role: role
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          role: role
          rel: service
          href: href
          type: application/json
          title: title
        keywords:
        - keywords
        - keywords
        description: description
        links:
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        id: id
        title: title
        version: version
        jobControlOptions:
        - null
        - null
      title: processSummary
    process:
      allOf:
      - $ref: '#/components/schemas/processSummary'
      - properties:
          inputs:
            additionalProperties:
              $ref: '#/components/schemas/inputDescription'
            type: object
          outputs:
            additionalProperties:
              $ref: '#/components/schemas/outputDescription'
            type: object
        type: object
      example:
        outputs:
          key:
            schema: null
            metadata:
            - null
            - null
            keywords:
            - keywords
            - keywords
            description: description
            title: title
        inputs:
          key:
            schema:
              $ref: $ref
            metadata:
            - hreflang: en
              role: role
              rel: service
              href: href
              type: application/json
              title: title
            - hreflang: en
              role: role
              rel: service
              href: href
              type: application/json
              title: title
            keywords:
            - keywords
            - keywords
            minOccurs: 0
            valuePassing:
            - byValue
            - byValue
            description: description
            maxOccurs: 6
            title: title
      title: process
...
    jobList:
      example:
        jobs:
        - exception:
            instance: instance
            detail: detail
            type: type
            title: title
            status: 0
          jobID: jobID
          processID: processID
          created: 2000-01-23T04:56:07.000+00:00
          progress: 8
          started: 2000-01-23T04:56:07.000+00:00
          finished: 2000-01-23T04:56:07.000+00:00
          links:
          - hreflang: en
            rel: service
            href: href
            type: application/json
            title: title
          - hreflang: en
            rel: service
            href: href
            type: application/json
            title: title
          type: process
          message: message
          updated: 2000-01-23T04:56:07.000+00:00
          status: null
        - exception:
            instance: instance
            detail: detail
            type: type
            title: title
            status: 0
          jobID: jobID
          processID: processID
          created: 2000-01-23T04:56:07.000+00:00
          progress: 8
          started: 2000-01-23T04:56:07.000+00:00
          finished: 2000-01-23T04:56:07.000+00:00
          links:
          - hreflang: en
            rel: service
            href: href
            type: application/json
            title: title
          - hreflang: en
            rel: service
            href: href
            type: application/json
            title: title
          type: process
          message: message
          updated: 2000-01-23T04:56:07.000+00:00
          status: null
        links:
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
      properties:
        jobs:
          items:
            $ref: '#/components/schemas/statusInfo'
          title: jobs
          type: array
        links:
          items:
            $ref: '#/components/schemas/link'
          title: links
          type: array
      required:
      - jobs
      - links
      title: jobList
      type: object
...
    descriptionType:
      properties:
        title:
          title: title
          type: string
        description:
          title: description
          type: string
        keywords:
          items:
            type: string
          title: keywords
          type: array
        metadata:
          items:
            $ref: '#/components/schemas/metadata'
          title: metadata
          type: array
      title: descriptionType
      type: object
    binaryInputValue:
      format: byte
      type: string
    execute:
      properties:
        inputs:
          additionalProperties:
            $ref: '#/components/schemas/input'
          title: inputs
          type: object
        outputs:
          additionalProperties:
            $ref: '#/components/schemas/output'
          title: outputs
          type: object
        subscriber:
          $ref: '#/components/schemas/subscriber'
      title: execute
      type: object
    format:
      properties:
        mediaType:
          title: mediaType
          type: string
        encoding:
          title: encoding
          type: string
        schema:
          $ref: '#/components/schemas/format_schema'
      title: format
      type: object
    inputDescription:
      allOf:
      - $ref: '#/components/schemas/descriptionType'
      - properties:
          valuePassing:
            default:
            - byValue
            - byReference
            items:
              enum:
              - byValue
              - byReference
              type: string
            type: array
          minOccurs:
            default: 1
            type: integer
          maxOccurs:
            $ref: '#/components/schemas/inputDescription_allOf_maxOccurs'
          schema:
            $ref: '#/components/schemas/schema'
        required:
        - schema
        type: object
      example:
        schema:
          $ref: $ref
        metadata:
        - hreflang: en
          role: role
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          role: role
          rel: service
          href: href
          type: application/json
          title: title
        keywords:
        - keywords
        - keywords
        minOccurs: 0
        valuePassing:
        - byValue
        - byValue
        description: description
        maxOccurs: 6
        title: title
      title: inputDescription
...
    inputValueNoObject:
      oneOf:
      - type: string
      - type: number
      - type: integer
      - type: boolean
      - items:
          type: object
        type: array
      - $ref: '#/components/schemas/binaryInputValue'
      - $ref: '#/components/schemas/bbox'
      title: inputValueNoObject
    jobControlOptions:
      enum:
      - sync-execute
      - async-execute
      - dismiss
      title: jobControlOptions
      type: string
    metadata:
      oneOf:
      - $ref: '#/components/schemas/metadata_oneOf'
      - $ref: '#/components/schemas/metadata_oneOf_1'
      title: metadata
    output:
      properties:
        format:
          $ref: '#/components/schemas/format'
      title: output
      type: object
    outputDescription:
      allOf:
      - $ref: '#/components/schemas/descriptionType'
      - properties:
          schema:
            $ref: '#/components/schemas/schema'
        required:
        - schema
        type: object
      example:
        schema: null
        metadata:
        - null
        - null
        keywords:
        - keywords
        - keywords
        description: description
        title: title
      title: outputDescription
    qualifiedInputValue:
      allOf:
      - $ref: '#/components/schemas/format'
      - properties:
          value:
            $ref: '#/components/schemas/inputValue'
        required:
        - value
        type: object
      title: qualifiedInputValue
...
    results:
      additionalProperties:
        $ref: '#/components/schemas/inlineOrRefData'
      type: object
    schema:
      oneOf:
      - $ref: '#/components/schemas/reference'
      - $ref: '#/components/schemas/schema_oneOf'
      title: schema
    statusCode:
      enum:
      - accepted
      - running
      - successful
      - failed
      - dismissed
      nullable: false
      title: statusCode
      type: string
    subscriber:
      description: |-
        Optional URIs for callbacks for this job.

        Support for this parameter is not required and the parameter may be
        removed from the API definition, if conformance class **'callback'**
        is not listed in the conformance declaration under `/conformance`.
      properties:
        successUri:
          format: uri
          title: successUri
          type: string
        inProgressUri:
          format: uri
          title: inProgressUri
          type: string
        failedUri:
          format: uri
          title: failedUri
          type: string
      required:
      - successUri
      title: subscriber
      type: object
    inlineOrRefData:
      oneOf:
      - $ref: '#/components/schemas/inputValueNoObject'
      - $ref: '#/components/schemas/qualifiedInputValue'
      - $ref: '#/components/schemas/link'
      title: inlineOrRefData
    statusInfo:
      example:
        exception:
          instance: instance
          detail: detail
          type: type
          title: title
          status: 0
        jobID: jobID
        processID: processID
        created: 2000-01-23T04:56:07.000+00:00
        progress: 8
        started: 2000-01-23T04:56:07.000+00:00
        finished: 2000-01-23T04:56:07.000+00:00
        links:
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        type: process
        message: message
        updated: 2000-01-23T04:56:07.000+00:00
        status: null
      properties:
        processID:
          title: processID
          type: string
        type:
          enum:
          - process
          title: type
          type: string
        jobID:
          title: jobID
          type: string
        status:
          $ref: '#/components/schemas/statusCode'
        message:
          title: message
          type: string
        exception:
          $ref: '#/components/schemas/exception'
        created:
          format: date-time
          title: created
          type: string
        started:
          format: date-time
          title: started
          type: string
        finished:
          format: date-time
          title: finished
          type: string
        updated:
          format: date-time
          title: updated
          type: string
        progress:
          maximum: 100
          minimum: 0
          title: progress
          type: integer
        links:
          items:
            $ref: '#/components/schemas/link'
          title: links
          type: array
      required:
      - jobID
      - status
      - type
      title: statusInfo
      type: object
...
    exception:
      additionalProperties: true
      description: JSON schema for exceptions based on RFC 7807
      example:
        instance: instance
        detail: detail
        type: type
        title: title
        status: 0
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required:
      - type
      title: Exception Schema
      type: object
...
    processes-list:
      enum:
      - RenderMap
      - ElevationContours
      - OSMERE
      title: processes-list
      type: string
    FeatureCollection:
      properties:
        type:
          enum:
          - FeatureCollection
          title: type
          type: string
        features:
          items:
            $ref: '#/components/schemas/GeoJSON_Feature'
          title: features
          type: array
        bbox:
          items:
            type: number
          minItems: 4
          title: bbox
          type: array
      required:
      - features
      - type
      title: GeoJSON FeatureCollection
      type: object
    results_1:
      additionalProperties:
        $ref: '#/components/schemas/inlineOrRefData'
      type: object
    link:
      example:
        hreflang: en
        rel: service
        href: href
        type: application/json
        title: title
      properties:
        href:
          title: href
          type: string
        rel:
          example: service
          title: rel
          type: string
        type:
          example: application/json
          title: type
          type: string
        hreflang:
          example: en
          title: hreflang
          type: string
        title:
          title: title
          type: string
      required:
      - href
      title: link
      type: object
...
    input:
      oneOf:
      - $ref: '#/components/schemas/inlineOrRefData_1'
      - items:
          $ref: '#/components/schemas/inlineOrRefData_1'
        type: array
      title: input
...
    inlineOrRefData_1:
      oneOf:
      - $ref: '#/components/schemas/inputValueNoObject_1'
      - $ref: '#/components/schemas/qualifiedInputValue_1'
      - $ref: '#/components/schemas/link'
      title: inlineOrRefData_1
...
    execute_200_response:
      oneOf:
      - type: string
      - type: number
      - type: integer
      - nullable: true
        type: object
      - items:
          type: object
        type: array
      - type: boolean
      - format: binary
        type: string
      - $ref: '#/components/schemas/results'
      title: execute_200_response
    execute_200_response_1:
      allOf:
      - format: geojson-feature-collection
        type: object
      - $ref: '#/components/schemas/FeatureCollection'
      title: execute_200_response_1
...
    inputDescription_allOf_maxOccurs:
      oneOf:
      - default: 1
        type: integer
      - enum:
        - unbounded
        type: string
      title: inputDescription_allOf_maxOccurs
...
    GeoJSON_Feature:
      properties:
        type:
          enum:
          - Feature
          title: type
          type: string
        id:
          $ref: '#/components/schemas/GeoJSON_Feature_id'
        properties:
          nullable: true
          title: properties
          type: object
        geometry:
          $ref: '#/components/schemas/GeoJSON_Feature_geometry'
        bbox:
          items:
            type: number
          minItems: 4
          title: bbox
          type: array
      required:
      - geometry
      - properties
      - type
      title: GeoJSON Feature
      type: object