# unity_sps_ogc_processes_api/src/unity_sps_ogc_processes_api/apis/processes_api.py
# coding: utf-8

from typing import Dict, List  # noqa: F401
import importlib
import pkgutil

from unity_sps_ogc_processes_api.apis.processes_api_base import BaseProcessesApi
import openapi_server.impl

from fastapi import (  # noqa: F401
    APIRouter,
    Body,
    Cookie,
    Depends,
    Form,
    Header,
    Path,
    Query,
    Response,
    Security,
    status,
)

from unity_sps_ogc_processes_api.models.extra_models import TokenModel  # noqa: F401
from unity_sps_ogc_processes_api.models.exception import Exception
from unity_sps_ogc_processes_api.models.execute import Execute
from unity_sps_ogc_processes_api.models.execute200_response import Execute200Response
from unity_sps_ogc_processes_api.models.execute200_response1 import Execute200Response1
from unity_sps_ogc_processes_api.models.process import Process
from unity_sps_ogc_processes_api.models.process_list import ProcessList
from unity_sps_ogc_processes_api.models.processes_list import ProcessesList
from unity_sps_ogc_processes_api.models.status_info import StatusInfo


router = APIRouter()

ns_pkg = openapi_server.impl
for _, name, _ in pkgutil.iter_modules(ns_pkg.__path__, ns_pkg.__name__ + "."):
    importlib.import_module(name)


@router.post(
    "/processes/{processId}/execution",
    responses={
        200: {"model": Execute200Response, "description": "Result of synchronous execution"},
        201: {"model": StatusInfo, "description": "Started asynchronous execution. Created job."},
        404: {"model": Exception, "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value."},
        500: {"model": Exception, "description": "A server error occurred."},
    },
    tags=["Processes"],
    summary="execute a process.",
    response_model_by_alias=True,
)
async def execute(
    processId:  = Path(..., description=""),
    execute: Execute = Body(None, description="An execution request specifying any inputs for the process to execute, and optionally to select specific outputs. With support for _Part 3: Workflows and chaining_, this execution request may specify a complex processing workflow e.g., including nested processes and OGC API collections as inputs."),
    prefer: str = Header(None, description="Indicates client preferences, including whether the client is capable of asynchronous processing. A &#x60;respond-async&#x60; preference indicates a preference for asynchronous processing. A &#x60;wait: &lt;x&gt;s&#x60; preference indicates that the client prefers to wait up to x seconds to receive a reponse synchronously before the server falls back to asynchronous processing."),
) -> Execute200Response:
    """Executes a process (this may result in the creation of a job resource e.g., for _asynchronous execution_).  For more information, see [Section 7.9](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_create_job). """
    return BaseProcessesApi.subclasses[0]().execute(processId, execute, prefer)


@router.get(
    "/processes/{processId}",
    responses={
        200: {"model": Process, "description": "A process description."},
        404: {"model": Exception, "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value."},
    },
    tags=["Processes"],
    summary="retrieve a process description",
    response_model_by_alias=True,
)
async def get_process_description(
    processId:  = Path(..., description=""),
) -> Process:
    """The process description contains information about inputs and outputs and a link to the execution-endpoint for the process. The Core does not mandate the use of a specific process description to specify the interface of a process. That said, the Core requirements class makes the following recommendation:  Implementations SHOULD consider supporting the OGC process description.  For more information, see [Section 7.8](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_description). """
    return BaseProcessesApi.subclasses[0]().get_process_description(processId)


@router.get(
    "/processes",
    responses={
        200: {"model": ProcessList, "description": "Information about the available processes"},
    },
    tags=["Processes"],
    summary="retrieve the list of available processes",
    response_model_by_alias=True,
)
async def get_processes(
) -> ProcessList:
    """The list of processes contains a summary of each process the OGC API - Processes offers, including the link to a more detailed description of the process.  For more information, see [Section 7.7]https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_list). """
    return BaseProcessesApi.subclasses[0]().get_processes()